+++ GuysNight.LethalCompanyMod.BalancedItems/Models/VanillaValues.cs
@@ -1,0 +1,23 @@
+﻿using System.Collections.Generic;
+
+namespace GuysNight.LethalCompanyMod.BalancedItems.Models {
+	public sealed class VanillaValues {
+		public VanillaValues(int minValue, int maxValue, float weight) {
+			MinValue = minValue;
+			MaxValue = maxValue;
+			Weight = weight;
+		}
+
+		public int MinValue { get; }
+
+		public int MaxValue { get; }
+
+		public Dictionary<string, byte> MoonRarities { get; } = new Dictionary<string, byte>();
+
+		public float Weight { get; }
+
+		public override string ToString() {
+			return $"MinValue: '{MinValue}' MaxValue: '{MaxValue}' Weight:'{Weight}'";
+		}
+	}
+}