+++ GuysNight.LethalCompanyMod.BalancedItems/Patches/GrabbableObjectPatches.cs
@@ -10,28 +10,41 @@ namespace GuysNight.LethalCompanyMod.BalancedItems.Patches {
 		[HarmonyPostfix]
 		public static void ChangeItemWeight(GrabbableObject __instance) {
 			if (__instance is null) {
-				SharedComponents.Logger.LogInfo($"__instance is null in '{nameof(ChangeItemWeight)}'. Aborting.");
+				SharedComponents.Logger.LogWarning($"__instance is null in '{nameof(ChangeItemWeight)}'. Aborting.");
 
 				return;
 			}
 
-			SharedComponents.Logger.LogInfo($"item.itemProperties.name is '{__instance.itemProperties.name}'");
-			SharedComponents.Logger.LogInfo($"item.itemProperties.weight is '{NumericUtilities.DenormalizeWeight(__instance.itemProperties.weight)}'");
-			SharedComponents.Logger.LogInfo($"item.itemProperties.minValue is '{__instance.itemProperties.minValue}'");
-			SharedComponents.Logger.LogInfo($"item.itemProperties.maxValue is '{__instance.itemProperties.maxValue}'");
+			SharedComponents.Logger.LogDebug($"item.itemProperties.name is '{__instance.itemProperties.name}'");
+			SharedComponents.Logger.LogDebug($"item.itemProperties.weight is '{NumericUtilities.DenormalizeWeight(__instance.itemProperties.weight)}'");
+			SharedComponents.Logger.LogDebug($"item.itemProperties.minValue is '{__instance.itemProperties.minValue}'");
+			SharedComponents.Logger.LogDebug($"item.itemProperties.maxValue is '{__instance.itemProperties.maxValue}'");
 
-			SharedComponents.Logger.LogInfo($"Begin adding config entries and setting override values for '{__instance.itemProperties.name}'");
+			SharedComponents.Logger.LogDebug($"Begin adding config entries and setting override values for '{__instance.itemProperties.name}'");
 
-			ConfigUtilities.SyncConfigForItemOverrides(__instance.itemProperties, out var itemOverrides);
+			var itemEntry = ConfigUtilities.SyncConfigForItemOverrides(__instance.itemProperties);
 
-			UpdateItemWeight(__instance, itemOverrides.Weight);
+			if (bool.TryParse(SharedComponents.ConfigFile[Constants.ConfigSectionHeaderToggles, Constants.ConfigKeyToggleWeights].GetSerializedValue(), out var isWeightFeatureEnabled)) {
+				SharedComponents.Logger.LogDebug($"Successfully retrieved weight override feature toggle. Value is '{isWeightFeatureEnabled}'");
+			}
+			else {
+				SharedComponents.Logger.LogWarning("Could not retrieve weight override feature toggle from config. Assuming it was set to true.");
+				isWeightFeatureEnabled = true;
+			}
+
+			if (isWeightFeatureEnabled) {
+				UpdateItemWeight(__instance, itemEntry.Overrides.Weight);
+			}
+			else {
+				UpdateItemWeight(__instance, itemEntry.VanillaValues.Weight);
+			}
 
 			SharedComponents.ConfigFile.Save();
 		}
 
 		private static void UpdateItemWeight(GrabbableObject item, float weight) {
 			item.itemProperties.weight = weight;
-			SharedComponents.Logger.LogInfo($"Successfully overrode weight for '{item.itemProperties.name}' to be '{NumericUtilities.DenormalizeWeight(weight)}'");
+			SharedComponents.Logger.LogInfo($"Successfully set weight for '{item.itemProperties.name}' to be '{NumericUtilities.DenormalizeWeight(weight)}'");
 		}
 	}
 }